requirements.txt
boto3
pg8000
python-dotenv
setuptools